Detect empty Github repositories
Don't point source_web_url to github if the repository created there is
empty.

Add explicit github_web_url and svn_web_url to the JSON data so that
presence of github repositories can be discovered, even if they are
empty.  Likewise, we may want to be able to distinguish packages that
have been moved to Github from packages that never existed in svn.